refactor(checker): split bare return call context helper
